Break up translations into separate directories
