+++ jsconfig.json
@@ -1,8 +1,11 @@
 {
   "compilerOptions": {
     "baseUrl": ".",
+    "module": "esnext",
+    "moduleResolution": "node",
     "paths": {
       "~/*": ["src/*"]
-    }
+    },
+    "types": ["vite/client"]
   }
 }
