+++ src/i18n/ar/translations.ts
@@ -1,0 +1,16 @@
+import { UIDictionary } from '../checks';
+
+export default UIDictionary({
+	'leftSidebar.learnTab': 'تَعلَم',
+	'leftSidebar.referenceTab': 'مرجع',
+	'rightSidebar.onThisPage': 'في الصفحة الحالية',
+	'rightSidebar.overview': 'نظرة عامة',
+	'rightSidebar.more': 'المزيد',
+	'rightSidebar.editPage': 'عدل هذه الصفحة',
+	'rightSidebar.translatePage': 'ترجم هذه الصفحة',
+	'rightSidebar.joinCommunity': 'انضم إلى مُجتمعنا',
+	'articleNav.nextPage': 'الصفحة التالية',
+	'articleNav.prevPage': 'عودة',
+	'since.addedIn': 'أُضيفت في:',
+	'since.new': 'جديدة',
+});
