+++ src/i18n/en/docsearch.ts
@@ -1,0 +1,8 @@
+import { DocSearchDictionary } from '../checks';
+
+export default DocSearchDictionary({
+	button: 'Search',
+	placeholder: 'Search docs',
+	shortcutLabel: 'Press / to search',
+	translations: {},
+});
