+++ src/i18n/pt-BR/translations.ts
@@ -1,0 +1,4 @@
+import { UIDictionary } from '../checks';
+
+export default UIDictionary({
+});
